+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-disable first-line-h1 line-length -->
 ## 4.6.0
 
-*2026-09-10*
+*2026-09-15*
 
 **Added**
 
@@ -9,6 +9,7 @@
 - The [autocompletion](./README.md#autocompletion) extension now also suggests root CSS variables in the CSS mode on a MediaWiki site
 - The [openLinks](./README.md#openlinks) extension now also supports Lua functions `mw.ext.TemplateStyles.link` on a MediaWiki site
 - The [autocompletion](./README.md#autocompletion) extension now also suggests [loadable libraries](https://www.mediawiki.org/wiki/Extension:Scribunto/Lua_reference_manual#Loadable_libraries) in the Lua mode on a MediaWiki site
+- The [openLinks](./README.md#openlinks) and [autocompletion](./README.md#autocompletion) extensions now also support localized aliases for the `Module` namespace in the Lua mode on a MediaWiki site
 
 ## 4.5.1